Changed Message client_time to Double.

+++ ChatRoom/Message.swift
@@ -8,21 +8,21 @@
 
 import Foundation
 
+// Message Class
 struct Message: Codable {
     var msg: String
-    var client_time:Int
+    var client_time:Double
     var server_time:Int?
     
-    init(msg: String, client_time: Int) {
+    init(msg: String, client_time: Double) {
         self.msg = msg
         self.client_time = client_time
         self.server_time = nil
     }
     
-    init(msg: String, client_time: Int, server_time: Int) {
+    init(msg: String, client_time: Double, server_time: Int) {
         self.msg = msg
         self.client_time = client_time
         self.server_time = server_time
     }
 }
-
